Move generally useful function to general helpers
