+++ tests/blueprints/test_cherrypicker_data_blueprint.py
@@ -5,8 +5,8 @@ from unittest.mock import patch
 import pytest
 
 from crawler.constants import FIELD_STATUS_COMPLETED, FLASK_ERROR_MISSING_PARAMETERS, FLASK_ERROR_UNEXPECTED
+from crawler.helpers.general_helpers import is_found_in_list
 from crawler.jobs.cherrypicker_test_data import CherrypickerDataError
-from tests.conftest import is_found_in_list
 
 barcode_metadata = [
     ["Plate-1", "positive samples: 30"],
